better SCA e2e tests
