+++ pkg/sca/testdata/libcap.yaml
@@ -28,40 +28,3 @@ pipeline:
       chmod 755 "${{targets.destdir}}/usr/lib/libcap.so.${{package.version}}"
 
   - uses: strip
-
-subpackages:
-  - name: libcap-dev
-    description: libcap development headers
-    pipeline:
-      - uses: split/dev
-    dependencies:
-      runtime:
-        - libcap
-        - libcap-utils
-
-  - name: libcap-utils
-    description: various utilities included with libcap
-    pipeline:
-      - runs: |
-          set -x
-          mkdir -p "${{targets.subpkgdir}}"/usr/sbin
-          for i in capsh getcap getpcaps setcap; do
-            mv "${{targets.destdir}}"/usr/sbin/$i "${{targets.subpkgdir}}"/usr/sbin/
-          done
-          rm -rf "${{targets.destdir}}/usr/sbin"
-    dependencies:
-      runtime:
-        - libcap
-
-  - name: libcap-doc
-    description: libcap documentation
-    pipeline:
-      - uses: split/manpages
-      - runs: |
-          set -x
-          rm -rf "${{targets.destdir}}/usr/share"
-
-update:
-  enabled: true
-  release-monitor:
-    identifier: 1569
